refactor: 재암호화 로직 libs로 분리
